% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduleDetail.R
\name{schedule_query_detail}
\alias{schedule_query_detail}
\title{查询数据}
\usage{
schedule_query_detail(
  token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3",
  FStartPlanDate = "2022-04-30",
  FEndPlanDate = "2022-05-20",
  FPrdNumber_start = "802685",
  FPrdNumber_end = "802685",
  FMoNumber_start = "MO202204499",
  FMoNumber_end = "MO202204499",
  FMachineNumber_start = "4#",
  FMachineNumber_end = "4#"
)
}
\arguments{
\item{token}{口令}

\item{FStartPlanDate}{计划日期}

\item{FEndPlanDate}{计划日期}

\item{FPrdNumber_start}{单据编码}

\item{FPrdNumber_end}{产品代码}

\item{FMoNumber_start}{开始}

\item{FMoNumber_end}{结束}

\item{FMachineNumber_start}{开始}

\item{FMachineNumber_end}{结束}
}
\value{
返回值
}
\description{
查询数据
}
\examples{
schedule_query_detail()
}
