% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduleStat.R
\name{schedule_query_stat}
\alias{schedule_query_stat}
\title{生产订单排程信息}
\usage{
schedule_query_stat(
  token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3",
  FPrdNumber_start = "801203-99",
  FPrdNumber_end = "801203-99",
  FMoNumber = "MO202204665",
  FCompanyName = "苏州赛普生物技术有限公司",
  FWorkShop = "总部生产部",
  FRouteName = "",
  FPrdSeries = ""
)
}
\arguments{
\item{token}{口令}

\item{FPrdNumber_start}{产品编码从}

\item{FPrdNumber_end}{产品编码到}

\item{FMoNumber}{订单编码}

\item{FCompanyName}{公司}

\item{FWorkShop}{生产车间}

\item{FRouteName}{工序}

\item{FPrdSeries}{产品系列}
}
\value{
返回值
}
\description{
生产订单排程信息
}
\examples{
schedule_query_stat()
}
