% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduleDetail.R
\name{schedule_read_group}
\alias{schedule_read_group}
\title{读取数据}
\usage{
schedule_read_group(
  token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3",
  file_name = "data-raw/生产订单排产信息表.xlsx"
)
}
\arguments{
\item{token}{口令}

\item{file_name}{文件名}
}
\value{
返回值
}
\description{
读取数据
}
\examples{
schedule_read_group()
}
