% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduleDetail.R
\name{scheduleDetailServer}
\alias{scheduleDetailServer}
\title{编码影响的查询函数}
\usage{
scheduleDetailServer(input, output, session, token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{token}{口令}
}
\value{
返回值
}
\description{
编码影响的查询函数
}
\examples{
scheduleDetailServer()
}
