% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduleStat.R
\name{scheduleStatServer}
\alias{scheduleStatServer}
\title{生产订单排产统计表}
\usage{
scheduleStatServer(input, output, session, token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{token}{口令}
}
\value{
返回值
}
\description{
生产订单排产统计表
}
\examples{
scheduleStatServer()
}
