% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule_read}
\alias{schedule_read}
\title{读取生产模板数据}
\usage{
schedule_read(
  file_name = "data-raw/生产订单排产信息表.xlsx",
  token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3"
)
}
\arguments{
\item{file_name}{文件名}
}
\value{
返回值
}
\description{
读取生产模板数据
}
\examples{
schedule_read()
}
